requirements.txt
llama-cpp-python
chromadb
sentence-transformers
langchain
PyPDF2
pdfplumber
python-docx
beir
numpy